huggingface-space/requirements.txt
# HuggingFace Space Requirements for dots.ocr
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
gradio>=4.0.0
pillow>=9.5.0
accelerate>=0.20.0
safetensors>=0.3.0
bitsandbytes>=0.41.0
scipy>=1.10.0
numpy>=1.24.0
huggingface-hub>=0.17.0